krishisarth/backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
influxdb-client==1.40.0
redis==5.0.1
celery[redis]==5.3.6
paho-mqtt==1.6.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
pydantic[email]==2.6.1
pydantic-settings==2.1.0
httpx==0.26.0
boto3==1.34.44
pandas==2.2.0
scikit-learn==1.4.1.post1
tensorflow==2.16.1
pytest==8.0.0
pytest-cov==4.1.0
pytest-asyncio==0.23.5
